The dll builds again against CommonLibSSE master.
  Made changes to the fla files to resolve dependencies

+++ src/Events.cpp
@@ -66,9 +66,9 @@ namespace Events
 		}
 
 		if (xHealth->health > healthMin) {
-			xHealth->health -= healthRelMax * 0.05;
-			if (xHealth->health < healthMin) {
-				xHealth->health = healthMin;
+			xHealth->health -= healthRelMax * 0.05f;
+			if (xHealth->health < static_cast<float>(healthMin)) {
+				xHealth->health = static_cast<float>(healthMin);
 			}
 		}
